libsel4arm-vmm: move generic gic code to header
Move functionality that is common to gicv2 and gicv3 drivers from vgic.c
to vgic.h so it can be used by the gicv3 driver.
