+++ libsel4arm-vmm/src/arch/arm/devices/vgic.h
@@ -11,9 +11,7 @@
  */
 #include "../../../vm.h"
 
-extern const struct device dev_vgic_dist;
 extern const struct device dev_vgic_vcpu;
 extern const struct device dev_vgic_cpu;
-
 
 int handle_vgic_maintenance(vm_t *vm, int idx);
